Add rust_spellfile and rust_spellsuggest crates
