+++ rust_spell/src/lib.rs
@@ -1,107 +1,13 @@
-use std::collections::HashMap;
 use std::ffi::{CStr, CString};
-use std::fs::File;
-use std::io::{BufRead, BufReader};
 use std::os::raw::{c_char, c_int, c_uchar};
 use std::sync::{Mutex, OnceLock};
 
-// Constants matching Vim's spell checking flags
+use rust_spellfile::{load_dict, Trie};
+use rust_spellsuggest::suggest;
+
 const WF_ONECAP: c_int = 0x02;
 const WF_ALLCAP: c_int = 0x04;
 const WF_KEEPCAP: c_int = 0x80;
-
-#[derive(Default)]
-struct Node {
-    children: HashMap<char, Node>,
-    is_word: bool,
-}
-
-#[derive(Default)]
-struct Trie {
-    root: Node,
-}
-
-impl Trie {
-    fn new() -> Self {
-        Self { root: Node::default() }
-    }
-
-    fn insert(&mut self, word: &str) {
-        let mut node = &mut self.root;
-        for ch in word.chars() {
-            node = node.children.entry(ch).or_default();
-        }
-        node.is_word = true;
-    }
-
-    fn collect(node: &Node, prefix: &mut String, out: &mut Vec<String>) {
-        if node.is_word {
-            out.push(prefix.clone());
-        }
-        for (ch, child) in &node.children {
-            prefix.push(*ch);
-            Self::collect(child, prefix, out);
-            prefix.pop();
-        }
-    }
-
-    fn all_words(&self) -> Vec<String> {
-        let mut out = Vec::new();
-        let mut prefix = String::new();
-        Self::collect(&self.root, &mut prefix, &mut out);
-        out
-    }
-
-    fn suggest(&self, word: &str, max: usize) -> Vec<String> {
-        let mut res = Vec::new();
-        for w in self.all_words() {
-            if edit_distance_one(word, &w) {
-                res.push(w);
-                if res.len() >= max {
-                    break;
-                }
-            }
-        }
-        res
-    }
-}
-
-fn edit_distance_one(a: &str, b: &str) -> bool {
-    if a == b {
-        return false;
-    }
-    let la = a.chars().count();
-    let lb = b.chars().count();
-    if la.abs_diff(lb) > 1 {
-        return false;
-    }
-    let a_chars: Vec<char> = a.chars().collect();
-    let b_chars: Vec<char> = b.chars().collect();
-    let mut i = 0usize;
-    let mut j = 0usize;
-    let mut diff = 0usize;
-    while i < la && j < lb {
-        if a_chars[i] == b_chars[j] {
-            i += 1;
-            j += 1;
-        } else {
-            diff += 1;
-            if diff > 1 {
-                return false;
-            }
-            if la > lb {
-                i += 1;
-            } else if lb > la {
-                j += 1;
-            } else {
-                i += 1;
-                j += 1;
-            }
-        }
-    }
-    diff += la - i + lb - j;
-    diff <= 1
-}
 
 static TRIE: OnceLock<Mutex<Trie>> = OnceLock::new();
 
@@ -115,39 +21,32 @@ pub extern "C" fn rs_spell_load_dict(path: *const c_char) -> bool {
         return false;
     }
     let cstr = unsafe { CStr::from_ptr(path) };
-    let Ok(p) = cstr.to_str() else { return false };
-    let file = match File::open(p) {
-        Ok(f) => f,
-        Err(_) => return false,
-    };
-    let reader = BufReader::new(file);
-    let mut trie = trie().lock().unwrap();
-    for line in reader.lines() {
-        if let Ok(word) = line {
-            let w = word.trim();
-            if !w.is_empty() {
-                trie.insert(w);
-            }
+    let Ok(p) = cstr.to_str() else { return false; };
+    match load_dict(p) {
+        Ok(t) => {
+            *trie().lock().unwrap() = t;
+            true
         }
+        Err(_) => false,
     }
-    true
 }
 
 #[no_mangle]
 pub extern "C" fn rs_spell_suggest(
     word: *const c_char,
     max: usize,
-    out_len: *mut usize,
+    len: *mut usize,
 ) -> *mut *mut c_char {
-    if word.is_null() || out_len.is_null() {
+    if word.is_null() || len.is_null() {
         return std::ptr::null_mut();
     }
     let cstr = unsafe { CStr::from_ptr(word) };
-    let Ok(w) = cstr.to_str() else { return std::ptr::null_mut() };
-    let trie = trie().lock().unwrap();
-    let suggestions = trie.suggest(w, max);
-    let len = suggestions.len();
-    unsafe { *out_len = len; }
+    let Ok(w) = cstr.to_str() else { return std::ptr::null_mut(); };
+    let suggestions = {
+        let trie_guard = trie().lock().unwrap();
+        suggest(&*trie_guard, w, max)
+    };
+    unsafe { *len = suggestions.len(); }
     let mut c_vec: Vec<*mut c_char> = suggestions
         .into_iter()
         .filter_map(|s| CString::new(s).ok().map(|cs| cs.into_raw()))
@@ -230,6 +129,7 @@ pub extern "C" fn captype(word: *const c_uchar, end: *const c_uchar) -> c_int {
 mod tests {
     use super::*;
     use std::ffi::{CStr, CString};
+    use std::fs::File;
     use std::io::Write;
     use std::ptr;
