+++ rust_spellfile/Cargo.toml
@@ -1,0 +1,6 @@
+[package]
+name = "rust_spellfile"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
