+++ rust_spellsuggest/Cargo.toml
@@ -1,0 +1,7 @@
+[package]
+name = "rust_spellsuggest"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+rust_spellfile = { path = "../rust_spellfile" }
