+++ src/proto.h
@@ -195,8 +195,6 @@ void mbyte_im_set_active(int active_arg);
 # endif
 # include "sound.pro"
 # include "spell.pro"
-# include "spellfile.pro"
-# include "spellsuggest.pro"
 # include "strings.pro"
 # include "syntax.pro"
 # include "tag.pro"
